// app/interfaces/MainComponent/OurProductComponent.ts
export interface IOurProduct {
    _id: string;
    id?: string;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
    }
    
    export interface IOurProductInput {
        id?: string;
        pagetitle?: string;
        title?: string;
        image?: string;
        description?: string;
    }